Extract parser token primitives
